+++ gwcEngine/gwcEngine/src/platform/OpenGL/OpenGLTime.cpp
@@ -9,7 +9,7 @@ namespace gwcEngine
 
 	void OpenGLTime::BeginFrameImpl()
 	{
-		m_TempTime = glfwGetTime();
+		m_TempTime = (float)glfwGetTime();
 		m_DeltaTime = m_TempTime-m_LastFrameTime;
 		m_LastFrameTime = m_TempTime;
 		m_Time += m_DeltaTime;
